refactor: 수강생 관리 관련 컴포넌트 파일명 수정
